refactor(bg-shell): split 1604-line god file into tool, command, and lifecycle modules
